增加歷程紀錄檔 新增卡片特效

+++ src/main/java/com/sample/controller/HelloController.java
@@ -11,7 +11,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.Part;
 
@@ -21,10 +20,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.sample.Util.ImageUtils;
+
 @Controller
 public class HelloController {
 
 	private String LOCALPATH = "D:\\workspace\\Springboot_MotoStore\\src\\main\\resources\\static\\image\\";
+	SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddHHmmssSSS");
 	@Autowired
 	private JdbcTemplate jdbcTemplate;
 
@@ -50,23 +52,29 @@ public class HelloController {
 	}
 
 	@RequestMapping("/insert")
-	public ModelAndView InsertData(HttpServletRequest request) throws IOException, ServletException {
-		SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddHHmmss");
-		String seqno = "T01" + sdf.format(new Date());
+	public ModelAndView InsertData(HttpServletRequest request) throws Exception {
+		
+		String seqnoT01 = "T01" + sdf.format(new Date());
+		String seqnoT02 = "T02" + sdf.format(new Date());
 		Part part = request.getPart("file");
 		String name = request.getParameter("name");
 		String price = request.getParameter("price");
 		String store = request.getParameter("store");
-		String sql = " INSERT INTO MT01 VALUES ('" + seqno + "','" + name + "','" + price + "','" + store + "','" + part.getSubmittedFileName() + "') ";
-		jdbcTemplate.update(sql);
+		Object[] parasT01 = new String[]{seqnoT01, name, price, store, part.getSubmittedFileName()};
+		Object[] parasT02 = new String[]{seqnoT02, seqnoT01, price, store, part.getSubmittedFileName(), name};
+		String sql = " INSERT INTO MT01 VALUES (?,?,?,?,?) ";
+		jdbcTemplate.update(sql, parasT01);
 		System.out.println(sql);
+		String sqlT2 =" INSERT INTO MT02 VALUES (?,?,?,?,?,?) ";
+		jdbcTemplate.update(sqlT2, parasT02);
+		System.out.println(sqlT2);
 		uploadPic(part);
 		ModelAndView mView = findAll();
 		return mView;
 	}
 
 	@RequestMapping("/update")
-	public ModelAndView UpdateData(HttpServletRequest request) throws IOException, ServletException {
+	public ModelAndView UpdateData(HttpServletRequest request) throws Exception {
 		String listCountString = request.getParameter("listCount");
 		int listCount = Integer.parseInt(listCountString);
 
@@ -74,17 +82,22 @@ public class HelloController {
 			String selchk = request.getParameter("selchk_" + i);
 
 			if ("Y".equals(selchk)) {
-				
+				String seqnoM2 = "T02" + sdf.format(new Date());
 				String id = request.getParameter("T01_ID_" + i);
 				String name = request.getParameter("T01_NAME_" + i);
 				String price = request.getParameter("T01_PRICE_" + i);
 				String store = request.getParameter("T01_STORE_" + i);
+				String picName = request.getParameter("T01_PICNAME_" + i);
 				Part part = request.getPart("T01_PIC_" + i);
-				String sql = " UPDATE MT01 SET T01_NAME='" + name + "',T01_PRICE='" + price + "',T01_STORE='" + store +
-						"',T01_PIC='" + part.getSubmittedFileName() + "' WHERE T01_ID='" + id + "' ";
-				jdbcTemplate.update(sql);
-				System.out.println(sql);
-				
+				Object[] parasT01 = new String[]{name, price, store, picName, id};
+				Object[] parasT02 = new String[]{seqnoM2, id, price, store, picName, name};
+				String sqlT1 = " UPDATE MT01 SET T01_NAME=?, T01_PRICE=?, T01_STORE=?, T01_PICNAME=? WHERE T01_ID=? ";
+				jdbcTemplate.update(sqlT1, parasT01);
+				System.out.println(sqlT1);
+				String sqlT2 =" INSERT INTO MT02 VALUES (?,?,?,?,?,?) ";
+				jdbcTemplate.update(sqlT2, parasT02);
+				System.out.println(sqlT2);
+				uploadPic(part);
  			}
 		}
 		ModelAndView mView = findAll();
@@ -92,7 +105,7 @@ public class HelloController {
 	}
 
 	@RequestMapping("/delete")
-	public ModelAndView DeleteData(HttpServletRequest request) throws IOException, ServletException {
+	public ModelAndView DeleteData(HttpServletRequest request) throws Exception {
 		String listCountString = request.getParameter("listCount");
 		int listCount = Integer.parseInt(listCountString);
 
@@ -123,11 +136,34 @@ public class HelloController {
 		return mView;
 	}
 	
-	public void uploadPic(Part part) throws IOException, ServletException {
-		// 要上傳的目標檔案存放路徑
+	@RequestMapping("/detail")
+	public ModelAndView showDetail(HttpServletRequest request) {
+		String t01id = request.getParameter("T01_CHOSENID");
+		ModelAndView mView = new ModelAndView();
+		String sql = " SELECT * FROM MT02 WHERE T02_T01ID = '"+ t01id +"' ORDER BY T02_ID ";
+		List<Map<String, Object>> list = jdbcTemplate.queryForList(sql);
+		mView.addObject("list", list);
+		mView.setViewName("detail");
+		return mView;
+	}
+	
+	public void uploadPic(Part part) throws Exception {
 		
+		String partFileName = part.getSubmittedFileName();
+		if("".equals(partFileName)){
+			return;
+		}
+		// 要上傳的目標檔案存放路徑
+		String[] nameparas = partFileName.split("\\.");
+		String name = nameparas[0];
+		String suffix = nameparas[1];
+		
+		String filepath = LOCALPATH + "Ori_" + partFileName;
+		if(new File(filepath).exists()) {
+			return;
+		}
 		try(InputStream in = part.getInputStream();  
-		OutputStream out = new FileOutputStream(LOCALPATH + part.getSubmittedFileName())) {
+		OutputStream out = new FileOutputStream(filepath)) {
 	        byte[] buffer = new byte[1024];
 	        int length = -1;
 	        while ((length = in.read(buffer)) != -1) {
@@ -136,13 +172,19 @@ public class HelloController {
 		} catch(IOException ex) {
 	            throw new UncheckedIOException(ex);
 	    }
+		ImageUtils iU = new ImageUtils(filepath, LOCALPATH, name, suffix);
+		iU.zoom(1920, 1080);
 	}
 	
-	public void delePic(String filename) throws IOException, ServletException {
-		String fileString = LOCALPATH + filename;
-		System.out.println(filename);
+	public void delePic(String filename) throws Exception {
 		
+		String fileString = LOCALPATH + filename;
+		String fileOriString = LOCALPATH + "Ori_" + filename;
 		File file = new File(fileString);
-		file.delete();
+		File fileOri = new File(fileOriString);
+		if(file.exists() && fileOri.exists()) {
+			file.delete();
+			fileOri.delete();
+		}
 	}
 }
